Separated sources in multiple files. Improved project structure
