+++ Sources/Widget/MultiHitsWidget.swift
@@ -1,0 +1,29 @@
+//
+//  MultiHitsWidget.swift
+//  InstantSearch
+//
+//  Created by Vladislav Fitc on 02/04/2019.
+//
+
+import Foundation
+import InstantSearchCore
+
+public protocol MultiHitsWidget: class {
+  
+  var viewModel: MultiHitsViewModel? { get set }
+  
+  func reload()
+  
+  func scrollToTop()
+  
+}
+
+public protocol MultiHitsSource: class {
+  
+  func numberOfSections() -> Int
+  func numberOfHits(inSection section: Int) -> Int
+  func hit<R: Codable>(atIndex index: Int, inSection section: Int) throws -> R?
+  
+}
+
+extension MultiHitsViewModel: MultiHitsSource {}
