+++ Sources/Widget/UIKit/Hits/CollectionViewHitsWidget.swift
@@ -10,9 +10,9 @@ import Foundation
 import InstantSearchCore
 import UIKit
 
-public typealias HitsCollectionViewController<Hit: Codable> = HitsController<CollectionViewHitsWidget<Hit>>
 public typealias CollectionViewCellConfigurator<Hit> = HitViewConfigurator<UICollectionView, UICollectionViewCell, Hit>
 public typealias CollectionViewClickHandler<Hit> = HitClickHandler<UICollectionView, Hit>
+public typealias CollectionViewHitsController<Hit: Codable> = HitsController<CollectionViewHitsWidget<Hit>>
 
 open class CollectionViewHitsDataSource<DataSource: HitsSource>: NSObject, UICollectionViewDataSource {
   
@@ -36,7 +36,7 @@ open class CollectionViewHitsDataSource<DataSource: HitsSource>: NSObject, UICol
   
 }
 
-open class HitsCollectionViewDelegate<DataSource: HitsSource>: NSObject, UICollectionViewDelegate {
+open class CollectionViewHitsDelegate<DataSource: HitsSource>: NSObject, UICollectionViewDelegate {
   
   public var clickHandler: CollectionViewClickHandler<DataSource.Record>
   public weak var hitsSource: DataSource?
@@ -57,7 +57,6 @@ open class HitsCollectionViewDelegate<DataSource: HitsSource>: NSObject, UIColle
 public class CollectionViewHitsWidget<Hit: Codable>: NSObject, HitsWidget {
   
   public typealias ViewModel = HitsViewModel<Hit>
-  public typealias SingleHitView = UICollectionViewCell
   
   public let collectionView: UICollectionView
   
@@ -70,7 +69,7 @@ public class CollectionViewHitsWidget<Hit: Codable>: NSObject, HitsWidget {
     }
   }
   
-  private var delegate: HitsCollectionViewDelegate<ViewModel>? {
+  private var delegate: CollectionViewHitsDelegate<ViewModel>? {
     didSet {
       delegate?.hitsSource = viewModel
       collectionView.delegate = delegate
