+++ Sources/Widget/UIKit/Hits/CollectionViewMultiHitsWidget.swift
