+++ Sources/Widget/UIKit/Hits/TableViewMultiHitsWidget.swift
